requirements.txt
streamlit
python-dotenv
langchain
langchain_community
langchain-groq
langchain-chroma
langchain-huggingface
Pypdf
sentence-transformers